refactor: new select verifible credential  ux
